Nonair/onair-minigame/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
google-generativeai
openai
jinja2
